+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Updated the credit card logo
+- API Keys are now stored with encryption provided by the Encryptor from the Magento Framework
 
 ## [2.12.2] - 2021-11-03
 ### Changed